+++ activerecord/test/cases/secure_token_test.rb
@@ -29,4 +29,34 @@ class SecureTokenTest < ActiveRecord::TestCase
 
     assert_equal @user.token, "custom-secure-token"
   end
+
+  def test_token_value_default_length_is_24
+    default_length = 24
+    @user.save
+    
+    assert_equal default_length, @user.token.length
+    assert_equal default_length, @user.auth_token.length
+    assert_not_equal default_length, @user.api_key.length
+  end
+
+  def test_token_length_option_change_token_length
+    @user.save
+    
+    assert_equal 42, @user.api_key.length
+  end
+  
+  def test_regenerating_token_with_same_length
+    @user.save
+    @user.regenerate_api_key
+    
+    assert_equal 24, @user.token.length
+    assert_equal 24, @user.auth_token.length
+    assert_equal 42, @user.api_key.length
+  end
+  
+  def test_prefix_prepend_in_token
+    @user.save
+    
+    assert @user.api_key.start_with?("ak_")
+  end
 end