+++ activerecord/test/schema/schema.rb
@@ -1009,6 +1009,7 @@ ActiveRecord::Schema.define do
   create_table :users, force: true do |t|
     t.string :token
     t.string :auth_token
+    t.string :api_key
   end
 
   create_table :test_with_keyword_column_name, force: true do |t|